+++ annotator/lineart_anime/__init__.py
@@ -142,7 +142,7 @@ class LineartAnimeDetector:
             image_feed = rearrange(image_feed, 'h w c -> 1 c h w')
 
             line = self.model(image_feed)[0, 0] * 127.5 + 127.5
-            line = line.cpu().numpy()
+            line = line.cpu().float().numpy()
 
             line = cv2.resize(line, (W, H), interpolation=cv2.INTER_CUBIC)
             line = line.clip(0, 255).astype(np.uint8)
